bug: 42318
Comcast does not set the feature to "active" just because both fields are set. However, WE remove both fields if the feature is inactive, which we probably shouldn't.

This update ensures that the fields are actually saved to comcast, even if the feature is inactive, for SCF and SCR.

http://bugzilla.zimbra.com/show_bug.cgi?id=42318

Copied from Perforce
 Change: 217250