bug: 42345
A check on newFeatures.getCallForwardingNoAnswer() != null caused an empty ZCallForwardingBean object to be created. This object was subsequently detected when submitting features to the backend, causing the behaviour in comment #2. When re-displaying the page, the dummy object is found, and it contains no "rings count" value, therefore the select box defaults to the first option (2).

http://bugzilla.zimbra.com/show_bug.cgi?id=42345

Copied from Perforce
 Change: 217258