bug 83440 -- IE11 support for IronMaiden
Loosely based on change 481303 but adapted to the IE compatibility
modes. Usually, the server will see the altered IE11 User Agent;
however, once we opt for quirks mode, IE11 will trigger it's
compatibility mode and appear as IE7 to the browser.

As compatibility mode -- and quirks mode in general -- is unlikely to
recognise IE10+ CSS gradients, we simplify the case in SkinResources
and always use ActiveX filters on IE.

http://bugzilla.zimbra.com/show_bug.cgi?id=83440

Copied from Perforce
 Change: 481505