bug 92861 - fix sending non-text messages in standard client
The standard client had a check for whether a message had any textual
contents, and if not, set the textual contents to the HTML source of
the message -- which seems a bit odd.

Instead, only set the textual content when specified, and likewise for
the textual content. As a result, a message with nothing but a smiley
won't specify any plain text fallback.

As best I can tell, this code has been in place since 2007 -- made in
bug 16825, comment 10 -- when Zimbra used a completely different
editor. In all probability, it's never worked with TinyMCE; however,
messages with no text whatsoever were either uncommon or impossible to
create until recently.

http://bugzilla.zimbra.com/show_bug.cgi?id=92861

Copied from Perforce
 Change: 516742