bug 92861 - properly fix sending non-text messages in standard client
In a cruel twist of fate, my previous fix in changelist 516742 was
doubly wrong: I fixed the wrong thing, and got the fix wrong.

Rather than only using the parts when they're non-empty, we'd only use
actually empty parts. However, that didn't affect the original bug,
which was that HTML-only emails actually inserted the textual message
as the HTML contents -- in a clause for handling the textual content
being empty.

http://bugzilla.zimbra.com/show_bug.cgi?id=92861

Copied from Perforce
 Change: 517160