Bug:97900
Html client uses fileAsStr string instead of fullName like the Ajax client uses when picking contacts/creating associated email address in compose window.
This causes an issue in Japanese as the "," character from fileAsStr's "Last, First" format isn't appropriate.

Decided not to make a Japanese specific change for the algorithm for generating fileAsStr as not sure what the consequences would be and also, if that was the fix, it wouldn't help with legacy data.
https://reviewboard.eng.zimbra.com/r/4259/

https://bugzilla.zimbra.com/show_bug.cgi?id=97900

Copied from Perforce
 Change: 538891