Bug 97900 - Standard HTML client inserts comma (,) between Family and Given name on compose, in Japanese preference [Problem Description] Html client uses fileAsStr string when picking contacts/creating associated email address in compose window.  But the fileAsStr did not comply with the locale specific label formatting pattern; it always used ',' as a delimiter if the FileAs was set to 'last first'.
Because the end user recognizes the contact item by the display label string, if the label string in address picker is different than the one in the Contact list, the end user may be confused whether the contact item which he/she selects during the message composing is the one which he/she created in the Contact list.  For the integrity of the UI, the label string for each contact item should be identical at any place, and among any clients.

[Root Cause]
The label string in the contact list was generated by the <app:contactFileAs> tag, but the one in the address picker was taken from the fileAsStr or later fullName.

At the first fix, the fileAsStr was replaced by the fullName.  Most case, fullName field contained a correct string complied with the FileAs value and formatting pattern.  But due to the old bug in the ZCS 5 or 6, the fullName might have a wrong value which ignored the fileAs value and the label formatting pattern, or even fullName was not set at all.  In that case, the fileAsStr was still used.  As a result, when the old contact item (for example, which was migrated from the older version) had a correct fullName value, the label string in the address picker window called from the compose window displayed the different label than the Contact list.

[Fix Description]
The address picker page uses the <app:contactFileAs> tag to generate the same display label string as the contact list.
If the contactHit is a personal contact item, the address picker page gets the lastName, firstName and other information field from the ZContactBean object, and passes them to the contactFileAs tag which computes the label string based on the FileAs value and localized formatting pattern defined at the ZmMsg.properties file.
If the contactHit is a group item, the address picker page resolves the members of the group, and extracts the lastName, firstName, etc. from each member, then passes them to the contactFileAs tag.

[Affected Area]
HTML Client: Compose window, address picker
Ajax client: None
Mobile client: None

[Configuration Changes]
None

[LDAP Schema Changes]
None

[API Changes]
None

[Log Changes]
None

[Manual Test]
(Locale: Japanese)
1/ Set the Web client language to Japanese
2/ Created a couple of contact items which the fullName and/or FileAs fields are missing (please refer to the sample contact items posted on the bug 98176, comment 9)
3/ Logined to the html client and opened the compose window, then opened the address picker window.
4/ Verified that the display label string of each contact item on the address picker list was identical to the one in the Contact list.

5/ Selected some contact item, set them to To/Cc/Bcc area, and moved back to the compose window.
6/ Verified that the display name string (shown next to the email address surrounded by double quotation marks) was same as the one in the address picker list, as well as the one in the Contact list.

7/ Created a contact group which included some contact items created at the Step 2.
8/ Opened the address picker window again
9/ Verified that the group name was shown identically to the Contact list.

10/ Selected the contact group, and moved back to the compose window.
11/ Verified that the display name for each member of the contact group was identical to the one in the Contact list.

(Locale: English)
Set the Web client language to English, and repeated the Step 2 to 11.

https://bugzilla.zimbra.com/show_bug.cgi?id=97900

Copied from Perforce
 Change: 541659