Bug:102684 - Composed message is not sent and it is lost sometimes
Desc:
This issue is seen because in "SendMsgRequest", the client sometimes sends a used "suid" to the server. 
As the server sees that this "suid" belongs to some older mail which was sent earlier, it sends a sucess response back. Due to this reason even though the client gets a positive "SendMsgResponse", the new mail is not sent at the backend.
To  fix this issue, we will always create  a new "suid" in the constructor of the ZMessageComposeBean and avoid it getting re-set.

Changelist:
1. ZMessageComposeBean: Set the "mSendUID" in the constructor to a random number. Have removed the setter for "mSendUID" field.
2. ZComposeUploaderBean: Have removed the code that  sets the value "mSendUID" because sometimes the value is stale.

https://bugzilla.zimbra.com/show_bug.cgi?id=102684

Copied from Perforce
 Change: 561995