bug: 103154 - df - clicking on compose button does not bring up the compose view
Renamed isDefault() in ZIdentity to getIsDefault(). This is a speculative fix based on the stack trace in the bug (& mocompose jsp has a reference to identity.isDefault). Fix can be validated post next dogfood push. The issue was most likely introduced post last jetty upgrade. In any case this looks like a right change to make.

https://bugzilla.zimbra.com/show_bug.cgi?id=103154

Copied from Perforce
 Change: 563178