ZBUG-696: Web-split:NPE thrown by unknown account at login
[Problem]
NullPointerException was thrown on the UI server side when logging in with a non-existing account on the multi-server web-split environment.  AS a result, "A client error occurred. Please try again." was shown in login page. Because this was caused by the lack of valid login ID, "The username or password is incorrect. Verify that CAPS LOCK is not on, and then retype the current username and password." error message should be displayed.

[Root cause]
When the environment is enabled to the WebClient split installation, the UI server submits to the RoutingLookupService running on the backend mailbox server.  The backend mailbox server looks up the account name to the LDAP and return the host info for the further SOAP request.　
The UI server creates a "server object" based on the response from the backend server; but the UI server didn’t take account into the fact that the “server object” could be null when the backend mailbox server didn’t find any info of the requested account.

[Fix]
Added the code to check whether the returned "server object" (nginxLookUpServer local variable) is null before accessing it.
(note) If the backend servers are all down, ServiceException is thrown and the error will be handled properly (that code has already worked as expected).

[Affected area]
Login tag for all client